plainComponentsのリファクタリング
