+++ lib/runtime/containers/EnqueteRuntimeApp.js
@@ -1,11 +1,15 @@
 /* eslint-env browser */
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
+import $ from 'jquery';
+import '../../jquery.plugins';
 import _ from 'underscore';
 import classNames from 'classnames';
+import 'photoswipe/dist/photoswipe.css';
+import 'photoswipe/dist/default-skin/default-skin.css';
 import Page from '../components/Page';
 import Finisher from '../components/Finisher';
-import plainComponents from '../../plainComponents';
+import { plainComponents } from '../components/plain/PlainComponents';
 import SurveyManager from '../SurveyManager';
 import * as Actions from '../actions';
 import { fixIphoneChromeBug } from '../../utils';
